requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dnspython==2.5.0
Flask==3.0.1
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
pymongo==4.6.1
python-dotenv==1.0.1
Werkzeug==3.0.1
zipp==3.17.0
